requirements.txt
discord.py
discord-pretty-help
discord.py[voice]
asyncio
lavalink
jishaku